+++ application/controllers/categories.php
@@ -5,6 +5,9 @@ class Categories_Controller extends Base_Controller {
 	// Домашний метод, выдает список категорий, кнопки управления
 	//--------------------------------------------------------------------------------------------------
 	public function action_index() {
+        if(!Auth::user())
+            return Redirect::to('login');
+
         $categories = Categories::getAllCategories();
 		$view = View::make('categories.home')->with('navActive', 'categories')->with('tree', $categories);
 
@@ -15,6 +18,9 @@ class Categories_Controller extends Base_Controller {
 	// Добавление категории
 	//--------------------------------------------------------------------------------------------------
 	public function action_addCategory() {
+        if(!Auth::user())
+            return Redirect::to('login');
+
 		$nameCategory = $_POST['nameCategory'];
         if(!empty($nameCategory)) {
             $parentId = 0;
@@ -27,6 +33,9 @@ class Categories_Controller extends Base_Controller {
     // Удаление категории
     //--------------------------------------------------------------------------------------------------
     public function action_deleteCategory() {
+        if(!Auth::user())
+            return Redirect::to('login');
+
         $idCategory = $_POST['idCategory'];
         if(!empty($idCategory)) {
             $category = Categories::find($idCategory)->delete();
@@ -38,6 +47,9 @@ class Categories_Controller extends Base_Controller {
     // Информация о категории
     //--------------------------------------------------------------------------------------------------
     public function getCategory($id) {
+        if(!Auth::user())
+            return Redirect::to('login');
+
         $idCategory = $_POST['idCategory'];
         if(!empty($idCategory)) {
 
@@ -48,6 +60,9 @@ class Categories_Controller extends Base_Controller {
     // Информация о категории в формате JSON
     //--------------------------------------------------------------------------------------------------
     public function action_getCategoryJson () {
+        if(!Auth::user())
+            return Redirect::to('login');
+
         $idCategory = $_POST['idCategory'];
         if(!empty($idCategory)) {
             $category = Categories::find($idCategory);
@@ -63,6 +78,9 @@ class Categories_Controller extends Base_Controller {
     // Строиит дерево категорий
     //--------------------------------------------------------------------------------------------------
     public  function action_buildCategoriesTree() {
+        if(!Auth::user())
+            return Redirect::to('login');
+
         $categories = Categories::getAllCategories();
         $view = View::make('categories.tree')->with('tree', $categories);
 
@@ -73,6 +91,9 @@ class Categories_Controller extends Base_Controller {
     // Обновление категории
     //--------------------------------------------------------------------------------------------------
     public function action_updateCategory() {
+        if(!Auth::user())
+            return Redirect::to('login');
+
         $idCategory   = $_POST['idCategory'];
         $nameCategory = $_POST['nameCategory'];
         if(!empty($idCategory) && !empty($nameCategory)) {
